Fix clang10 warnings
Summary:
- Clang 10 has a new warning: "-Wimplicit-int-float-conversion" which
  flags some code. Fix these up by explicitly casting where appropriate.
- Clang 10 uses "-Wdefaulted-function-deleted" which warns about
  `ExecutorTimeoutManager` special member functions being implicitly
  deleted due to the base class of `TimeoutManager` having a deleted move
  constructor. Instead of defaulting them when they will be deleted,
  just remove them.

+++ folly/ConcurrentSkipList-inl.h
@@ -236,7 +236,7 @@ class SkipListRandomHeight {
       p *= kProb;
       sizeLimit *= kProbInv;
       lookupTable_[i] = lookupTable_[i - 1] + p;
-      sizeLimitTable_[i] = sizeLimit > kMaxSizeLimit
+      sizeLimitTable_[i] = sizeLimit > static_cast<double>(kMaxSizeLimit)
           ? kMaxSizeLimit
           : static_cast<size_t>(sizeLimit);
     }
